マーカー通過時に距離補正が働かない問題を修正

+++ robotrace_v2/Core/Inc/control.h
@@ -84,7 +84,7 @@ extern speedParam tgtParam;
 // マーカー関連
 extern uint8_t courseMarker;
 extern uint8_t beforeCourseMarker;
-extern uint32_t cntMarker;
+extern uint16_t cntMarker;
 extern uint8_t courseMarkerLog;
 extern int32_t straightMeter;
 extern bool straightState;
